Misc updates to the WGS pipelines (#302)
* Separate small, SV call, and:

  * FixSnifflesVCF guard against when only a subset of issues exist in the VCF

  * Drop Clair3 from CLR data, as that's not supported

* FOUR SMALL FIXES:

  * guard against spaces in LB field in input BAM....

  * optionally check PacBio ShardLongreads outputs (sometimes EOF mark missing)

  * use DV 1.3.0, bundled in custom docker with resource monitoring script

  * improved PBSV resource allocation based on scatter scheme choice,
    as PAPI 10 observed for high coverage data when not proecssing per chr

* provide default values to WGS pipelines

+++ wdl/tasks/VariantUtils.wdl
@@ -64,17 +64,22 @@ task MergeAndSortVCFs {
 
     input {
         Array[File] vcfs
-        Array[File] tbis
 
         File ref_fasta_fai
+        File? header_definitions_file
 
         String prefix
 
         RuntimeAttr? runtime_attr_override
     }
+    parameter_meta {
+        header_definitions_file: "a union of definition header lines for input VCFs (related to https://github.com/samtools/bcftools/issues/1629)"
+    }
 
     Int sz = ceil(size(vcfs, 'GB'))
     Int disk_sz = if sz > 100 then 5 * sz else 375  # it's rare to see such large gVCFs, for now
+
+    Boolean suspected_incomplete_definitions = defined(header_definitions_file)
 
     Int cores = 8
 
@@ -105,11 +110,42 @@ task MergeAndSortVCFs {
         echo "==========================================================="
         echo "done concatenation, fixing header of naively concatenated VCF" && date
         echo "==========================================================="
+        if ~{suspected_incomplete_definitions}; then
+            # a bug from bcftools concat --naive https://github.com/samtools/bcftools/issues/1629
+            set +e
+            zgrep "^##" concatedated_raw.vcf.gz > header.txt
+            grep -vF 'fileformat' header.txt \
+                | grep -vF 'fileDate=' \
+                | grep -vF 'source=' \
+                | grep -vF 'contig' \
+                | grep -vF 'ALT' \
+                | grep -vF 'FILTER' \
+                | grep -vF 'INFO' \
+                | grep -vF 'FORMAT' \
+                > tmp.others.txt
+            touch tmp.other.txt
+            set -e
+            zgrep "^#CHROM" concatedated_raw.vcf.gz > tmp.sampleline.txt
+            cat \
+                ~{header_definitions_file} \
+                tmp.others.txt \
+                tmp.sampleline.txt \
+                > fixed.header.txt
+            rm -f tmp.*.txt && cat fixed.header.txt
+
+            bcftools reheader \
+                -h fixed.header.txt \
+                -o tmp.wgs.vcf.gz \
+                concatedated_raw.vcf.gz
+            rm concatedated_raw.vcf.gz
+        else
+            mv concatedated_raw.vcf.gz tmp.wgs.vcf.gz
+        fi
         bcftools reheader \
             --fai ~{ref_fasta_fai} \
             -o wgs_raw.vcf.gz \
-            concatedated_raw.vcf.gz
-        rm concatedated_raw.vcf.gz
+            tmp.wgs.vcf.gz
+        rm tmp.wgs.vcf.gz
 
         echo "==========================================================="
         echo "starting sort operation" && date
@@ -135,6 +171,71 @@ task MergeAndSortVCFs {
     RuntimeAttr default_attr = object {
         cpu_cores:          cores,
         mem_gb:             "~{machine_memory}",
+        disk_gb:            disk_sz,
+        boot_disk_gb:       10,
+        preemptible_tries:  1,
+        max_retries:        0,
+        docker:             "us.gcr.io/broad-dsp-lrma/lr-basic:latest"
+    }
+    RuntimeAttr runtime_attr = select_first([runtime_attr_override, default_attr])
+    runtime {
+        cpu:                    select_first([runtime_attr.cpu_cores,         default_attr.cpu_cores])
+        memory:                 select_first([runtime_attr.mem_gb,            default_attr.mem_gb]) + " GiB"
+        disks: "local-disk " +  select_first([runtime_attr.disk_gb,           default_attr.disk_gb]) + " LOCAL"
+        bootDiskSizeGb:         select_first([runtime_attr.boot_disk_gb,      default_attr.boot_disk_gb])
+        preemptible:            select_first([runtime_attr.preemptible_tries, default_attr.preemptible_tries])
+        maxRetries:             select_first([runtime_attr.max_retries,       default_attr.max_retries])
+        docker:                 select_first([runtime_attr.docker,            default_attr.docker])
+    }
+}
+
+task CollectDefinitions {
+    meta {
+        description: "Collect (union) various definitions in vcf files, adddressing a bcftols bug: https://github.com/samtools/bcftools/issues/1629"
+    }
+    input {
+        Array[File] vcfs
+
+        RuntimeAttr? runtime_attr_override
+    }
+
+    Int sz = ceil(size(vcfs, 'GB'))
+    Int disk_sz = if sz > 100 then 5 * sz else 375
+
+    command <<<
+        set -euxo pipefail
+
+        zgrep "^##" ~{vcfs[0]} > header.txt
+        grep -F '##fileformat' header.txt > tmp.0.txt
+        grep -F '##fileDate=' header.txt > tmp.1.txt
+        if grep -q -F '##source=' header.txt; then grep -F 'source=' header.txt > tmp.2.txt; fi
+        touch tmp.2.txt
+        grep -F '##contig=' header.txt > tmp.3.txt
+
+        cat tmp*txt > easy.txt && rm tmp*txt
+
+        touch tmp.alt.txt tmp.ft.txt tmp.info.txt tmp.format.txt
+        for vcf in ~{sep=' ' vcfs}; do
+            zgrep -F '##ALT=' "${vcf}" >> tmp.alt.txt
+            zgrep -F '##FILTER=' "${vcf}" >> tmp.ft.txt
+            zgrep -F '##INFO=' "${vcf}" >> tmp.info.txt
+            zgrep -F '##FORMAT=' "${vcf}" >> tmp.format.txt
+        done
+        for txt in tmp*txt; do
+            sort "${txt}" | uniq > "${txt}.union"
+        done
+        cat tmp.alt.txt.union tmp.ft.txt.union tmp.info.txt.union tmp.format.txt.union > hard.txt
+        cat easy.txt hard.txt > definitions.union.txt
+    >>>
+
+    output {
+        File union_definitions = "definitions.union.txt"
+    }
+
+    #########################
+    RuntimeAttr default_attr = object {
+        cpu_cores:          4,
+        mem_gb:             16,
         disk_gb:            disk_sz,
         boot_disk_gb:       10,
         preemptible_tries:  1,
@@ -366,12 +467,14 @@ task FixSnifflesVCF {
     command <<<
         set -euxo pipefail
 
+        # 1. fix sample information (Sniffles derives VCF SM information from the path to the BAM ......)
         cp ~{vcf} ~{local_raw}
         echo ~{sample_name} > sample_names.txt
         bcftools reheader --samples sample_names.txt -o ~{local_sp_fixed} ~{local_raw}
         rm ~{vcf} && rm ~{local_raw}
 
         ####################################################################
+        # 2. prep for fixing undefined VCF INFO/FT/FORMAT, also guard against when the VCF is empty
         ~{initial_grep_cmd} "^##" ~{local_sp_fixed} > header.txt
         ~{initial_grep_cmd} -v "^#" ~{local_sp_fixed} > body.txt || true
         if [[ ! -f body.txt ]] || [[ ! -s body.txt ]]; then
@@ -387,25 +490,53 @@ task FixSnifflesVCF {
         fi
 
         ####################################################################
+        # 2.1. more prep for fixing undefined VCF INFO/FT/FORMATs
         # get FORMATs in header
-        grep -F '##FORMAT=<' header.txt | awk -F ',' '{print $1}' | sed 's/##FORMAT=<ID=//' | sort > formats.in_header.txt
-        # get FILTERs in practice
-        grep -F '##FILTER=<' header.txt | awk -F ',' '{print $1}' | sed 's/##FILTER=<ID=//' | sort > filters.in_header.txt
-        # get non-flag INFO in practice
-        grep -F '##INFO=<' header.txt | grep -vF 'Type=Flag' | awk -F ',' '{print $1}' | sed 's/##INFO=<ID=//' | sort > non_flag_info.in_header.txt
-        # get     flag INFO in practice
-        grep -F '##INFO=<' header.txt | grep  -F 'Type=Flag' | awk -F ',' '{print $1}' | sed 's/##INFO=<ID=//' | sort >     flag_info.in_header.txt
+        if grep -q -F '##FORMAT=<' header.txt; then
+            grep -F '##FORMAT=<' header.txt | awk -F ',' '{print $1}' | sed 's/##FORMAT=<ID=//' | sort > formats.in_header.txt
+        else
+            touch formats.in_header.txt
+        fi
+        # get FILTERs in header
+        if grep -q -F '##FILTER=<' header.txt; then
+            grep -F '##FILTER=<' header.txt | awk -F ',' '{print $1}' | sed 's/##FILTER=<ID=//' | sort > filters.in_header.txt
+        else
+            touch filters.in_header.txt
+        fi
+        # get non-flag INFO in header
+        if grep -q -F '##INFO=<' header.txt; then
+            grep -F '##INFO=<' header.txt | grep -vF 'Type=Flag' | awk -F ',' '{print $1}' | sed 's/##INFO=<ID=//' | sort > non_flag_info.in_header.txt
+        else
+            touch non_flag_info.in_header.txt
+        fi
+        # get     flag INFO in header
+        if grep -q -F '##INFO=<' header.txt; then
+            grep -F '##INFO=<' header.txt | grep  -F 'Type=Flag' | awk -F ',' '{print $1}' | sed 's/##INFO=<ID=//' | sort >     flag_info.in_header.txt
+        else
+            touch flag_info.in_header.txt
+        fi
 
         # get FORMATs in practice
         awk '{print $9}' body.txt | sort | uniq | sed 's/:/\n/g' | sort | uniq > formats.in_vcf.txt
-        # get FILTERs in practice
-        awk '{print $7}' body.txt | sort | uniq | grep -v "^PASS$" > filters.in_vcf.txt
-        # get non-flag INFO in practice
-        awk '{print $8}' body.txt | sed 's/;/\n/g' | grep -F '=' | awk -F '=' '{print $1}' | sort | uniq > non_flag_info.in_vcf.txt
-        # get     flag INFO in practice
-        awk '{print $8}' body.txt | sed 's/;/\n/g' | grep -vF '=' | sort | uniq > flag_info.in_vcf.txt
+        # get FILTERs in practice, guard against no 'PASS'
+        awk '{print $7}' body.txt | sort | uniq | grep -v "^PASS$" > filters.in_vcf.txt || touch filters.in_vcf.txt
+
+        awk '{print $8}' body.txt | sed 's/;/\n/g' > tmp.info.entries.txt
+        if grep -q -F '=' tmp.info.entries.txt; then
+            # get non-flag INFO in practicez
+            grep -F '=' tmp.info.entries.txt | awk -F '=' '{print $1}' | sort | uniq > non_flag_info.in_vcf.txt
+        fi
+        if grep -q -vF '=' tmp.info.entries.txt; then
+            # get     flag INFO in practice
+            awk '{print $8}' body.txt | sed 's/;/\n/g' | grep -vF '=' | sort | uniq > flag_info.in_vcf.txt
+        fi
+        touch non_flag_info.in_vcf.txt
+        touch     flag_info.in_vcf.txt
+
+        echo "I survived. More to go..."
 
         ####################################################################
+        # 2.2. more prep for fixing undefined VCF INFO/FT/FORMATs
         comm -13 formats.in_header.txt formats.in_vcf.txt > missing.formats.txt
         while IFS= read -r line
         do
@@ -431,20 +562,25 @@ task FixSnifflesVCF {
         done < missing.flag_info.txt
 
         ####################################################################
-        grep "^##" ~{local_sp_fixed} | grep -v "^##[A-Z]" | grep -vF 'contig=' > first_lines.txt
-        grep -F "##contig=<ID=" header.txt > contigs.txt
-        grep "^#CHROM" ~{local_sp_fixed} > sample.line.txt
-        grep "^##" ~{local_sp_fixed} | grep "^##[A-Z]" | sort > definitions.txt
-        touch missing.placeholder.header && \
-          cat definitions.txt missing.*.header | sort > everything.defined.txt
-        cat first_lines.txt contigs.txt everything.defined.txt sample.line.txt > fixed.header.txt
+        # 2. actually fix undefined VCF INFO/FT/FORMATs, if necessary
+        if  find . -maxdepth 1 -type f -name "missing.*.header" 2>/dev/null | grep -q .; then
+            grep "^##" ~{local_sp_fixed} | grep -v "^##[A-Z]" | grep -vF 'contig=' > first_lines.txt
+            grep -F "##contig=<ID=" header.txt > contigs.txt
+            grep "^#CHROM" ~{local_sp_fixed} > sample.line.txt
+            grep "^##" ~{local_sp_fixed} | grep "^##[A-Z]" | sort > existing_definitions.txt
+            cat existing_definitions.txt missing.*.header | sort > everything.defined.txt
+            cat first_lines.txt contigs.txt everything.defined.txt sample.line.txt > fixed.header.txt
+            # print to stdout for checking
+            grep -vF "##contig=<ID=" fixed.header.txt
 
-        # print to stdout for checking
-        grep -vF "##contig=<ID=" fixed.header.txt
+            cat fixed.header.txt body.txt > fixed.vcf
+            rm ~{local_sp_fixed}
+        else
+            mv ~{local_sp_fixed} fixed.vcf
+        fi
 
-        cat fixed.header.txt body.txt > fixed.vcf
-        rm ~{local_sp_fixed}
-
+        ####################################################################
+        # 3. fix contigs undefined (in later stages)
         if ~{fix_contigs}; then
             bcftools reheader \
                 --fai ~{ref_fasta_fai} \
@@ -453,6 +589,8 @@ task FixSnifflesVCF {
             mv fixed.and_contigs.vcf fixed.vcf
         fi
 
+        ####################################################################
+        # 4. fix occationally unsorted VCF
         bcftools \
             sort \
             --temp-dir tm_sort \
